v3.0.1
Fixed the bug that some Paper server failed to load.
Fixed the bug that the method of getting the number of pages in the mailing list in MailManager can only get one page.
Fixed the bug that the receiveAllPersonMail method in MailManager stops when a received email is encountered.
The send method is added to SystemMail, which will correctly trigger the MailboxSystemMailSendEvent event.